split classes into multiple files, added doc blocks
This makes it easier to maintain. A build script to create a single file
again could be added if wanted.
